refactor: properly organize environment configuration tests
- Move environment model tests from auth_service_test.go to dedicated env_config_test.go
- Remove unused imports from auth_service_test.go
- Maintain proper test file organization with related tests grouped together

This improves code organization and makes tests easier to find and maintain.
